+++ src/extension/config.ts
@@ -54,6 +54,10 @@ export function parseConfiguration(cfg: vscode.WorkspaceConfiguration): Extensio
                 cfg.get('linter.validateNamesRule.enabled') ?? DEFAULT_T.enableValidateNamesRule,
             enableKeywordCasingRule:
                 cfg.get('linter.keywordCasingRule.enabled') ?? DEFAULT_T.enableValidateStringRule,
+            checkReturnTypes:
+                cfg.get('linter.checkReturnTypes.enabled') ?? DEFAULT_A.checkReturnTypes,
+            checkUnusedLocals:
+                cfg.get('linter.checkUnusedLocals.enabled') ?? DEFAULT_A.checkUnusedLocals,
         }
     };
 }